+++ Tests/Wolverine.AspNetCore.Controllers/Wolverine.AspNetCore.Controllers.Infrastructure/Dispatch/Middleware/LoggingMiddleware.cs
@@ -23,6 +23,10 @@ namespace Wolverine.AspNetCore.Controllers.Infrastructure.Dispatch.Middleware
             ICurrentUserService currentUserService,
             CancellationToken cancellationToken)
         {
+            if (envelope.Message is null)
+            {
+                return;
+            }
             await LogAsync(envelope.Message, logger, currentUserService);
         }